# Optimizing conv kernels a bit