--- a/Cargo.toml
+++ b/Cargo.toml
@@ -32,7 +32,7 @@ matrixmultiply = { version = "0.3.2", default-features = false, optional = true 
 zip = { version = "0.6.2", default-features = false, optional = true }
 cblas-sys = { version = "0.1.4", default-features = false, optional = true }
 libc = { version = "0.2", default-features = false, optional = true }
-cudarc = { path = "../cudarc", default-features = false, optional = true, features = ["driver", "cublas"] }
+cudarc = { version = "0.9.2", default-features = false, optional = true, features = ["driver", "cublas"] }
 num-traits = { version = "0.2.15", default-features = false }
 safetensors = { version = "0.3", default-features = false, optional = true }
 memmap2 = { version = "0.5", default-features = false, optional = true }
@@ -46,7 +46,7 @@ indicatif = "0.17.3"
 glob = { version = "0.3.1", optional = true }
 
 [features]
-default = ["std", "fast-alloc", "cpu-par-matmul", "cuda", "nightly"]
+default = ["std", "fast-alloc", "cpu-par-matmul"]
 nightly = []
 
 std = ["cudarc?/std"]