--- a/src/tensor_ops/conv2d/cuda_kernel.rs
+++ b/src/tensor_ops/conv2d/cuda_kernel.rs
@@ -65,7 +65,7 @@ where
         let patches_item_numel = op.chan_in * op.kernel * op.kernel * op.h_out * op.w_out;
         let patches_numel = op.batch * patches_item_numel;
 
-        let mut patches = self.get_workspace::<E>(patches_numel)?;
+        let mut patches = unsafe { self.get_workspace::<E>(patches_numel) }?;
         let mut patches = unsafe { patches.transmute_mut::<E>(patches_numel).unwrap() };
 
         let img_strides = self.dev.htod_copy(make_4d::<L>(lhs.strides).into())?;
@@ -110,7 +110,7 @@ where
         let patches_numel = op.batch * patches_item_numel;
         let filters_numel = op.batch * op.chan_in * op.chan_out * op.kernel * op.kernel;
 
-        let mut patches = self.get_workspace::<E>(patches_numel)?;
+        let mut patches = unsafe { self.get_workspace::<E>(patches_numel) }?;
         let mut patches = unsafe { patches.transmute_mut::<E>(patches_numel).unwrap() };
 
         let mut f_b1023 = unsafe { self.dev.alloc::<E>(filters_numel) }?;