Resolves #547 
Related to #578 

This does a couple things:
1. Adds a workspace to Cuda that allows conv kernel to not have to re-allocate memory for patches
2. Removes a memset(0) from the output of Conv operation
3. Removes a memset(0) on patches allocation
4. No longer broadcasts filters in transpose_and_broadcast_filters
5. Sets `patches[i] = 0.0` for both unfold_input & unfold_output
6. Uses the parallel stream to parallelize conv operations a bit

Timings of `conv2d` bench on A10
| branch | forward | backward |
| --- | --- | --- |
| main | 9ms | 52ms |
| updates | 5ms | 21.5ms |